This TidyModule is used to explore the content of the ModStore.

Store is a TidyModule that can be used in your application to list existing applications, sessions and display your session's modules and edges.

Details

Should be initialized and injected in your application.

Super class

tidymodules::TidyModule -> Store

Methods

Public methods

Store$ui()

Store$server()

Store$clone()

Inherited methods

tidymodules::TidyModule$addInputPort()

tidymodules::TidyModule$addOutputPort()

tidymodules::TidyModule$assignPort()

tidymodules::TidyModule$callModule()

tidymodules::TidyModule$countInputPort()

tidymodules::TidyModule$countOutputPort()

tidymodules::TidyModule$deepClone()

tidymodules::TidyModule$definePort()

tidymodules::TidyModule$destroy()

tidymodules::TidyModule$doServer()

tidymodules::TidyModule$execInput()

tidymodules::TidyModule$execOutput()

tidymodules::TidyModule$getGlobalSession()

tidymodules::TidyModule$getInput()

tidymodules::TidyModule$getInputPort()

tidymodules::TidyModule$getInputPorts()

tidymodules::TidyModule$getOutput()

tidymodules::TidyModule$getOutputPort()

tidymodules::TidyModule$getOutputPorts()

tidymodules::TidyModule$getPortDef()

tidymodules::TidyModule$getPortName()

tidymodules::TidyModule$getSession()

tidymodules::TidyModule$getSessionId()

tidymodules::TidyModule$getShinyInput()

tidymodules::TidyModule$getShinyOutput()

tidymodules::TidyModule$getShinySession()

tidymodules::TidyModule$getStore()

tidymodules::TidyModule$initialize()

tidymodules::TidyModule$iport()

tidymodules::TidyModule$isGlobal()

tidymodules::TidyModule$isStored()

tidymodules::TidyModule$ns()

tidymodules::TidyModule$oport()

tidymodules::TidyModule$print()

tidymodules::TidyModule$render()

tidymodules::TidyModule$reset()

tidymodules::TidyModule$resume()

tidymodules::TidyModule$suspend()

tidymodules::TidyModule$updateInputPort()

tidymodules::TidyModule$updateInputPorts()

tidymodules::TidyModule$updateOutputPort()

tidymodules::TidyModule$updateOutputPorts()

tidymodules::TidyModule$use()

tidymodules::TidyModule$view()

Method ui()

Store's ui function.

Usage

Store$ui()

Returns

UI elements.

Method server()

Store's server function.

Usage

Store$server(input, output, session)

Arguments

input

Shiny input.

output

Shiny output

session

Shiny session

Method clone()

The objects of this class are cloneable with this method.

Usage

Store$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.