TidyModule R6 Class, parent of all modules

This is the main class provided by tidymodules and should be the parent of all tidymodules modules. It provides methods for defining input and output communication ports, managing module namespace and many other useful features.

Public fields

id

ID of the module.

module_ns

Module namespace, unique identifier for the module.

parent_ns

Parent module namespace in case of nested modules.

parent_mod

Reference to parent module in case of nested modules.

parent_ports

logical value indicating if the module inherit the parent's input ports in case of nested modules.

group

Group name of the module.

created

Initialization time of the module.

Methods

Public methods

TidyModule$new()

TidyModule$ns()

TidyModule$getSessionId()

TidyModule$render()

TidyModule$ui()

TidyModule$server()

TidyModule$definePort()

TidyModule$assignPort()

TidyModule$addInputPort()

TidyModule$updateInputPort()

TidyModule$updateInputPorts()

TidyModule$getInputPort()

TidyModule$getInputPorts()

TidyModule$getInput()

TidyModule$updateOutputPort()

TidyModule$updateOutputPorts()

TidyModule$addOutputPort()

TidyModule$getPortName()

TidyModule$getOutputPort()

TidyModule$getOutputPorts()

TidyModule$getOutput()

TidyModule$getStore()

TidyModule$countInputPort()

TidyModule$countOutputPort()

TidyModule$use()

TidyModule$callModule()

TidyModule$isStored()

TidyModule$isGlobal()

TidyModule$getSession()

TidyModule$getGlobalSession()

TidyModule$doServer()

TidyModule$getPortDef()

TidyModule$print()

TidyModule$deepClone()

TidyModule$reset()

TidyModule$getShinyInput()

TidyModule$getShinyOutput()

TidyModule$getShinySession()

TidyModule$clone()

Method new()

Create a new tidymodules module.

Usage

TidyModule$new(id = NULL, inherit = TRUE, group = NULL)

Arguments

id

Unique Id to assign to the module. Default to a generated Id using module's class and order of initialization.

inherit

logical value indicating if a nested module should inherits the parent's input ports. Default to TRUE

group

Module group name. Added to module's Id to make it unique. Optional

Returns

A new TidyModule object.

Method ns()

namespace function used to generate unique Id for HTML elements.

Usage

TidyModule$ns(id)

Arguments

id

Id of HTML element / shiny input.

Returns

A unique string Id.

Method getSessionId()

Get module session Id. This function rely on a shiny output object to find the right session Id.

Usage

TidyModule$getSessionId()

Returns

The Session Id of the module.

Method render()

Alias to the ui function.

Usage

TidyModule$render(...)

Arguments

...

arguments passed to the ui function.

Method ui()

UI function generating the module HTML elements. This function is eventually implemented in the child module to give the module a visual representation. Please note that a module can have many visual representations.

Usage

TidyModule$ui(...)

Arguments

...

arguments passed to ui

Returns

Empty tagList()

Method server()

server function to be overwritten and called by child module.

Usage

TidyModule$server(input, output, session)

Arguments

input

shiny input.

output

shiny output.

session

shiny session.

Method definePort()

Function wrapper for port definition expression.

Usage

TidyModule$definePort(x)

Arguments

x

expression

Method assignPort()

Function wrapper for port assignement expression.

Usage

TidyModule$assignPort(x)

Arguments

x

expression

Method addInputPort()

Add input port function. To be called within definePort function

Usage

TidyModule$addInputPort(
  name = NULL,
  description = NULL,
  sample = NULL,
  input = FALSE,
  is_parent = FALSE
)

Arguments

name

port name. must be a unique input port name.

description

port description.

sample

sample dataset consumed by this port. Mandatory!

input

This argument should be ignored. Only here for backward compatibility.

is_parent

Is the port inherited from the parent module.

Method updateInputPort()

Function for filling an input port. Called within the assignPort function

Usage

TidyModule$updateInputPort(id = NULL, input = NULL)

Arguments

id

Port name or Id .

input

The reacive object.

Method updateInputPorts()

This function add a set of input ports to the module.

Usage

TidyModule$updateInputPorts(inputs = NULL)

Arguments

inputs

reactivevalues with the input ports.

Method getInputPort()

Get an input port from the module.

Usage

TidyModule$getInputPort(id = 1)

Arguments

id

Name or Id of the port.

Returns

A module input port. A reactivevalues object with name, description, sample, is_parent and port elements.

Method getInputPorts()

Get all the input ports as a reactivevalues object.

Usage

TidyModule$getInputPorts()

Returns

A reactivevalues object.

Method getInput()

Get a input port slot.

Usage

TidyModule$getInput(id = 1)

Arguments

id

Name or Id of the port.

Returns

A reactive function or reactivevalues object.

Method updateOutputPort()

Function for filling an output port. Called within the assignPort function

Usage

TidyModule$updateOutputPort(id = NULL, output = NULL)

Arguments

id

Port name or Id .

output

The reacive object.

Method updateOutputPorts()

This function add a set of output ports to the module.

Usage

TidyModule$updateOutputPorts(outputs = NULL)

Arguments

outputs

reactivevalues with the output ports.

Method addOutputPort()

Add output port function. To be called within definePort function

Usage

TidyModule$addOutputPort(
  name = NULL,
  description = NULL,
  sample = NULL,
  output = FALSE,
  is_parent = FALSE
)

Arguments

name

port name. must be a unique output port name.

description

port description.

sample

sample dataset returned by this port. Mandatory!

output

This argument should be ignored. Only here for backward compatibility.

is_parent

Is the port inherited from the parent module.

Method getPortName()

Utility function that returns a port name from the Id.

Usage

TidyModule$getPortName(id = NULL, type = "input")

Arguments

id

Port Id

type

Port type, input or output.

Returns

Port name.

Method getOutputPort()

Get an output port from the module.

Usage

TidyModule$getOutputPort(id = 1)

Arguments

id

Name or Id of the port.

Returns

A module output port. A reactivevalues object with name, description, sample, is_parent and port elements.

Method getOutputPorts()

Get all the output ports as a reactivevalues object.

Usage

TidyModule$getOutputPorts()

Returns

A reactivevalues object.

Method getOutput()

Get a output port slot.

Usage

TidyModule$getOutput(id = 1)

Arguments

id

Name or Id of the port.

Returns

A reactive function or reactivevalues object.

Method getStore()

Function for retrieving the central ModStore.

Usage

TidyModule$getStore()

Returns

The ModStore object.

Method countInputPort()

Utility function that counts the number of input ports.

Usage

TidyModule$countInputPort()

Returns

The input ports count.

Method countOutputPort()

Utility function that counts the number of output ports.

Usage

TidyModule$countOutputPort()

Returns

The output ports count.

Method use()

Alias to the callModule function.

Usage

TidyModule$use(...)

Arguments

...

arguments passed to the server function of the module.

Method callModule()

callModule function. Similar to shiny callModule. Used to inject the server logic into the application. This function don't need the user to provide a namespace Id as a module already knows its identity. Options provided as arguments will be passed to the server function of the module. Note that the module reference self might not be the one injected.

Usage

TidyModule$callModule(...)

Arguments

...

arguments passed to the server function of the module.

Method isStored()

Function to check if the module is store in the current session.

Usage

TidyModule$isStored()

Returns

logical value.

Method isGlobal()

Check if the session attached to the module is the global_session.

Usage

TidyModule$isGlobal()

Returns

logical value.

Method getSession()

Get the current session.

Usage

TidyModule$getSession()

Returns

A reactivevalues object with the following elements. aid = application Id path = application path sid = session Id count = current module count created = session creation time updated = session update time collection = list of session modules edges = list of connection edges

Method getGlobalSession()

Get the global session.

Usage

TidyModule$getGlobalSession()

Returns

A reactivevalues object with the following elements. aid = application Id path = application path sid = "global_session" count = session module count created = session creation time updated = session update time collection = list of session modules edges = empty data.frame

Method doServer()

Function interfacing with shiny's callModule.

Usage

TidyModule$doServer(...)

Arguments

...

arguments passed to the server function of the module.

Method getPortDef()

Utility function to retrieve a port definition in the form of a list. This is a useful function to learn about a specific port.

Usage

TidyModule$getPortDef(type = NULL, id = 1)

Arguments

type

Port type, input or output.

id

Name or Id of port.

Returns

List of the port definition.

Method print()

Module printing function. Print the structure of a module.

Usage

TidyModule$print()

Examples

MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
m <- MyModule$new()
m

Method deepClone()

Module cloning function. Take care of ports (cloning reactive objects) and nested modules. Note that the Ids/namespace are not changed.

Usage

TidyModule$deepClone(o = NULL, i = NULL, s = NULL)

Arguments

o

Optional shiny output.

i

Optional shiny input

s

Optional shiny input

Returns

A cloned module.

Method reset()

This function reset the ports.

Usage

TidyModule$reset(o = NULL, i = NULL, s = NULL)

Arguments

o

Optional shiny output.

i

Optional shiny input

s

Optional shiny input

Method getShinyInput()

Retrieve the shiny input.

Usage

TidyModule$getShinyInput()

Returns

Shiny input object.

Method getShinyOutput()

Retrieve the shiny output.

Usage

TidyModule$getShinyOutput()

Returns

Shiny output object.

Method getShinySession()

Retrieve the shiny output.

Usage

TidyModule$getShinySession()

Returns

Shiny session object.

Method clone()

The objects of this class are cloneable with this method.

Usage

TidyModule$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `TidyModule$print`
## ------------------------------------------------

MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
m <- MyModule$new()
m

#> Module Namespace MyModule-3
#> Module Session global_session
#> - Class MyModule << TidyModule << R6
#> - Input [0]
#> - Output [0]